// types.ts
export enum EducationalStage {
  PRE_SCHOOL = 'PRE_SCHOOL',
  PRIMARY = 'PRIMARY',
  MIDDLE = 'MIDDLE',
  HIGH = 'HIGH',
}

export enum Subject {
  // Primary
  Arabic = 'اللغة العربية',
  French = 'اللغة الفرنسية',
  Math = 'الرياضيات',
  IslamicEducation = 'التربية الإسلامية',
  SocialStudies = 'الاجتماعيات',

  // Middle & High
  English = 'اللغة الإنجليزية',
  Physics = 'الفيزياء والكيمياء',
  NaturalSciences = 'علوم الحياة والأرض',
  ComputerScience = 'المعلوميات',
  Philosophy = 'الفلسفة',
}
// ...
export interface Grade {
    subSubject: string;
    semester: 1 | 2;
    assignment: 1 | 2;
    score: number | null;
}

export interface Student {
    id: string;
    guardianCode: string;
    name: string;
    stage: EducationalStage;
    level: string;
    class: string;
    grades: {
        [key in Subject]?: Grade[];
    };
}